ADD Peano project ExaHyPeDataStructure
Added the definition files for an ExaHyPe dummy project that
will create data structures for multiple PDEs with different
order of approximation on the same grid

Note that the code does not compile at this stage since
we need to add manual glue code in order to get the
Patch-based solver working. This will be done
in one of the next commits.