ExaHyPeDataStructure: CellDescription: Implemented M. Dumbser's edits
Implemented M. Dumbser's edits. See Thread
"ExaHyPE data structure" on the ExaHyPE mailing
list for more details.

The DoF on the faces are now stored in 2*d smaller arrays
instead of a single large one. Therefore, the user is not
required to stride from face to face as in the previous version.