NEW BRANCH: AdvectionDG0: Started development of Peano-based DG method
Started the development of a Peano-based DG method.
We will consider regular meshes and only
serial computations. Information regarding the DG method
can be found in subfolder "AdvectionDG0/share/15_peano_for_dg/"
You need to compile the corresponding tex document
with pdflatex.

NOTE: The code is not working yet.
I moved it into a separate branch in
order to continue working on it from home.