EulerFlow3d: More refinement levels; smaller patches
Changed the definitions in Constants.h to have
a default setup that better exploits Peano's
shared memory strategy.

The number of initial leaf cells is now (3^(l-1))^2=81,
with l=3 denoting the inital global refinement level.
Per leaf cell, we employ  3^2 patches.

We set the finest granularity to be a leaf cell.
Thus, we have enough work for 60+ Threads.