EulerFlow3d: Replaced 4 adapters (Predictor,FaceDataExchange,Corrector,GlobalTimeStep) by 2
This was made possible by reordering and combining adapters, and additionally
storing the time step size used in the previous iteration.

We now have the following two adapters expressed in terms of the old adapters.
  Adapter 1: Corrector(old dt)->Predictor(current dt)->GlobalTimeStepComputation(current dt)
  Adapter 2: ExchangeFaceData(current dt)
Here, "old dt" denotes the time step size in the old iteration n-1,
and "current dt" denotes the time step size in the current iteration n.

A more detailed explanation of the work that is done in Adapter 1:
  1. Perform the corrector step using the old update and the old global time step size.
     This is a leaf-cell-local operation. Thus we immediately obtain the leaf-cell-local current solution.
  2. Perform the predictor step using the leaf-cell-local current solution and the current global time step size.
  3. Compute the leaf-cell-local time step sizes
  4. (Optionally) Export the leaf-cell-local current solution.
  5. After the traversal, set the global current time step size as the new old global time step size.
     Find the minimum leaf-cell-local time step size and set it as the new current
     global time step size.

Tested with GCC and Intel TBB in Release mode.

NOTE: The idea for the above reordering and recombination of adapters stemmed from Tobias Weinzierl.
I do not take credit for it. I am just the guy who implemented it ;)