EulerFlow3d: Removed initial corrector sub step
The error in the previous commit could be traced back to a division
by zero in the PDE eigenvalue computation invoked by the Riemann solver.
The error occured since we initialised the extrapolated predictor
values to zero.

Since the only purpose of the initial corrector sub step is to
add a zero update to the initial solution, I decided
to introduce an additional adapter that does not
perform the corrector sub step.

Note that this adapter is called only once before the
time stepping. We still need only two grid traversals per
time step.

I tested the code on a system with gcc (serial & TBB).
After this commit, I will pull the code on another machine
with Intel compiler and test it there again.