Code/Python: Added code for analysing Peano Log output
The code comes with a small library 'runtimeparser.py'
that reads in user and CPU time of Peano adapters
from a Peano log output file.
This might be useful for monitoring the performance
of our applications after applying changes to
the kernels/the PDE engine.

The frontend script 'usertimeplot.py" creates
a speedup plot based on a a set of
Peano log output files with specific
file name pattern.

A documentation can be created using make.
For more details see the 'README.txt'.