ExaHyPE: trivialproject is now compiling
Please be aware that it does not run yet and that
it does not compute anything useful since I commented out
a lot of code in mappings/ExchangeFaceData and
mappings/RiemannSolver.
These two mappings need to be modified
and combined since we do not want to use
ghost layers anymore.