exahype & kernels: Mismatched space time predictor signatures caused seg fault
Unified the signatures and applied corresponding
changes to Toolkit.

Status quo:
Now the code runs into an invalid value assertion
in the plotter code. The solution blows up.
The first time step seems reasonable but the
consequent time steps do not differ from it.
This might indicate that the time step size
is not updated correctly.