New branch: Divided solver functionality into Solver and Solver class.
WARNING: Code is compiling and running in serial but results are wrong.
That is why I push my changes to a new branch.

The new concept is:
A cell descriptions links to a new Solve object which links
to one of the registered solvers.
The Solve objects are stored within the State and
must be hold consistent in shared memory (and MPI) mode.

Why do I do this?
The Solves will act as colouring, i.e., they will
allow us to identify runs that use the same kernels
but different boundary and initial conditions.
This might be useful for optimisation and uncertainty
identification.
If we further store the Solves in a tree-like structure,
we can add and remove Solves on-the-fly.
This might be useful for uncertainty identification.