Euler 3D Fortran: split space-time predictor
Following Dominic's idea, the space-time predictor has been split
into 3 subroutines, namely the picard loop, the integration in time,
and the extrapolation. In spaceTimePredictor.cpph, these routines
are called subsequently.

We revive this idea because linear and nonlinear numerics share parts
of the space-time predictor. And it's easier for debugging
my optimisation stuff.

spaceTimePredictor.cpph: delete[] reintroduced. Runs with the Intel
compiler, but in the event of a segmentation fault, you might want
to remove this again.