Improved `.gitignore` to ignore more Peano-related files (coming from extracting the tarball). Improved README.md file to contain a mini copy-and-paste guide to quickly install and compile ExaHyPE on most (linux) machines.
The Peano integration #34 is better, but still not optimal. Why do we need this extra step of keeping a tarball in the repository which has to be extracted after downloading the code (cloning the repository)? This always creates a big mess in the local git repository (for instance, the `peano.tar.gz` overwrites the `.gitignore` in the same directory).