CodeGen: adapt weight
As agreed in the last phone call, we want to move the quadrature
nodes/weights into the corresponding folder generic/ and
optimised/, respectively. I added an aligned and padded version
of the weights for my optimised version to no longer rely on
the generic version.

Updated the documentation (Guidebook) accordingly.