idealMHDfortran: Added the new Application. However, it is not finished and therefore also not yet tested (likely not to compile). We are missing the PDE source terms, this problem was raised 3 months ago in #21.
Also, there will be a lot of redundancy as I have to copy 90% of the srhd3dfortran code for the idealMHDfortran application. Instead, this really should be one single application where one can choose about what one wants to compile in the exahype specification file or by environment variables on compile time.