Fortran SRHD application: Basically merged Svens srhdhybrid and Vascos SRHD application which did the same: A C++ kernel which calls Fortran functions. This will be the basis for the new Fortran support in ExaHyPE. Once this is done, all srhd applications can be deleted in favour to this single implementation. Especially the buggy srhd2dcpp application can be deleted at this future date.
The SRHD application compiles, but for some reason the Makesystem does not compile the intermediate module typesDef.f90. I don't know why, but you can copy the generated typesDef.mod from srhd3dfortran for the time being. Currently, there are three modules: The eos in PDE.f90, the constants.f90 and the typesDef.f90. Of course we don't need all of them, but in general the Make system should just do it's job and compile all the f90 files -- in right order, which is currently probably not supported yet.

After compilation, the code crashes with "stack smashing detected". So: Yes, it compiles, but it doesn't run. Probably this is related to my switching back from the flux(Q,F) signature to flux(F,Q). Does it matter?