CodeGen: finalise SolutionUpdateGenerator
Follow the ideas of last jour fixe and use the
same structure as in the generic kernels for
luh. Rationale is export functionality (vtk,...)
and easier verification of results across
several iteration.

This version is slow. The compiler vectorises
it only to a limited degree. Later on we should
use the old version again, which is fully
vectorised by the compiler, but uses aos format.
This version is now commented out.