CodeGen: finalise StableTimeStepGenerator
I commented out the old version which
assumes soa format of luh, lduh. Following
the idea of the last jour fixe we use aos
format for the solution to be able to
compare results with the generic code base easily.

The implementation is ugly, but it's working.
And anyway it's only a temporary solution if
we later on return to soa.