SRHD now runs without crashes in 3D, 2D and with order=3, order=1.
The crash was produced by a wrong argument order when calling Fortran functions from C++. We should pay attention to this, as this is now different in the branch "SRHD-New-Fortrankernel" (ie. commit c16d035ce3b721a051ed4e67c147ce21af6cdc4c) and in the master (ie. this commit).