SRHD: Deleted the old srhd applications in favour for the new one.
The srhd2dcpp was a port of the srdh3dfortran to Cpp and contained bugs which were never solved. It will not be maintained any more. The srhd3dfortran was the initial application which used the old Fortran interface of ExaHyPE which will also pass away. We stick to the new Fortran interface with the new SRHD application which works (will work) in 2D and 3D.

If you want to work with the old applications, just checkout the last commit to a new branch.