Toolkit: [Bugfix] compare content of strings, not addresses
if (variable == "value")
does not compare the content of the strings but the addresses.
Replacing a switch handler with this if construct effectively
removed the ALIGNMENT flag written to the generated Makefile.

@all Please check in the toolkit if your string comparison
have also fallen victim to the downgrade of the Java version
(where switch cannot be used to compare strings)