ADERDGSolver: Two changes.
-Physics based refinement can now veto virtual refinement event.
-We further replace now virtual Descendant cells by real
cells (Cell) if the parent cell (Cell) had requested
refinement.

Solver/FiniteVolumesSolver/ADERDGSolver:

-Introduce a function hasToSendDataToMaster() which
tells a master if a worker rank has to send (non time step)
data to its master. This decision is made on the master and is
purely based on the cell type of the cell at the
master worker boundary. Is the cell
an ancestor of other real cells than
the worker will restrict up data
to this cell and send it up to the
master too afterwards.