Dear Z4 formulation of Einsteins Equations, welcome to ExaHyPE.
I adapted Olindos Trento fortran research code to the ExaHyPE infrastructure. I put this application to the folder "Applications/" and not to the folder "ApplicationExamples/" because I was told that the later is supposed to be open-sourced soon and we probably want to keep Einstein Equations related applications closed for some years.

The application in this commit will most likely not compile. It will definetly not work because it needs PDENCP kernel calls (non-conservative products) in the non linear ADER-DG kernels which are not yet present.