Removed all Application compile scripts. They now reside in Code/Misc/BuildScripts.
In contrast to the copies in the ApplicationExamples, there is only a single copy of the same compile script which is continously updated. You can make softlinks from the Application directories if you want to have the same setup as before (just being able to execute "./compile.sh"). You can also put the compile.sh to your PATH or use the exa command line utility.