lscompact: A tool for listing large directories in a compact manner.
With ExaHyPE we quickly get a directory with hundreds of VTK files from a single plotter, with six plotters an ordinary directory listing prevents us from seeing anything. This python script collects similar names in a fashion like "output[0-100].py", thus allowing to quickly inspect the contents of a directory. It especially supports complex ranges like "output[0-10,15,20-300].vtk". It's not very luxury but a first attemp of a really neccessary tool.