ADERDGCellDescription.cpp: Store additional heap indices for previous solution. We need this field for curing NaNs in limiter as well as for merging the limiting into the fused time stepping algorithm.
In the fused time stepping algorithm, the update array
is cleared and overwritten by volume integral contributions.
It is thus not possible anymore to rollback to a previous
solution by means of the update vector.

I plan to hold a copy of the previous solution,
and to initialise the current solution array with these values.
I then add all volume and surfacec integrals contributions directly to the current
solution array.
Thus, a rollback to the previous solution is
always possible even in the presence of NaNs in the current solution.

Consequently, I will kick out the update field in the near future.
I will first make it a temporary variable and keep
the kernel. We can then later on decide if we want to
merge the update kernel funcitonality into the other kernels.