Convergence tests: Base class for SRMHD tests. This is the way to go.
To be honest, this massive python starting file approach looks like a workaround for the failing cli argument system. But on the other hand, the nested classes offer new possibilities to simplify things. I think this is not bad at all and the ParametricTest class was a superfluous mistake where I'm too vain to delete it.