Include human/machine readable build and toolkit information into the binary. This greatly simplifies simulation managament.
This commit is a first try to extract runtime information from the toolkit down to the executable with command line arguments (--version).

A similar approach for actual build time information (environment variables in the Makefile) has to be reactivated (ExaHyPE/buildinfo.h).

This addresses #91 but not fully solves it as there are still crucial information missing.