EulerFlow: Moved parameter detection to constructor (init) for initial data and prepared blocks for run-time chosen boundary conditions.
This version is using a function pointer for the initial data. This is also more performant than the lookup in the environment variable dictionary. The former InitialData() function (now idfunc) is called at many places, so this probably even makes a difference in performance.

However it was actually motivated to implement similar techniques for the boundary conditions.