Applications, TimeSeriesReductions: Built in a poor man's MPI support, ie. not reducing over all ranks but letting only the first rank to do the writing out.
This fits well for the needs in issue #103, 2a), ie. I can read the plotting times for very long applications as I can run them with MPI and don't have to wait so long for the results. As this is only for measuring the plotting times and I'm not interested in the plotted data at all, this is fine.