Toolkit: Passed the cmdlineargs for the FV solvers.
Note that there is currently quite an inconsistency with auxilliary arguments passed to the solver constructors. For instance, for the ADERDG Generic solvers we do have user parameters passed while we don't have for the FV. At least most kernels now have the cmdline args.

This might break the ADERDG coupled to FV kernels as I haven't checked yet. But it's an easy fix.