SRMHD: Adapting Fortran signature for incontinous Flux storage (F matrix to Fx,Fy,Fz).
It compiles, but I could not test it as the overall ExaHyPE currently does not compile for me.

It should be checked that the AlfenWave still runs fine after this change.

The change also introduces some small performance improvements for the Fortran code, for instance the scattering of Q(var) to the local variables which is now handled by the preprocessor with ifdefs.
Also the Eigenvalue Magnetosonics are completely skipped in favour of L=DivCleaning_a=1.