EulerFlow: Replacing PastaMatrix linear algebra with tarch::la::Vector.
This smallish self-written vector library in it's nonrealtivistic use had no advantage over tarch::la. EulerFlow is supposed to be a simple to use demonstrator application so this is the right step.

Note: Have not tested if the initial data "MovingGauss2D" and "DiffusingGauss" are still correct. But if not, problems should be easy to find as this is initial data.