BuildScripts/ClusterConfig: Allow the make file to read machine configuration from the ClusterConfigs script.
Now it's just typing "exa compile EulerFlow" on my cluster and it goes, loads the correct compilers, runs the java toolkit, runs make -j and so on.

Have I already mentioned that updating ExaHyPE with this tool is as easy as "exa update all" which first makes a git pull and then in parallel a svn update peano and builds the toolkit.

And all that works from any directory.