GRMHD: Finished a first version of an integrator on a sphere.
It produces results, but I'm not sure whether they are right.

It's also not performant at all as it checks for all points on the sphere whether they are inside the cell which has to be plotted. This can be done in a more intelligent way.

I think this is a good starting point for Alejandro.