ADERDG 3D generic C kernels: Combined source function which gets gradQ in order to compute NCP itself. Also in Riemannsolver removed MatrixB in favour of NCP.
With this commit, the signature of the algebraicSource (acutally a composited source) changes to (Q,gradQ,S). The matrixB function is no more needed.

The change at the riemannSolverNonlinear is crucial for the CCZ4 application (it does no more have a dedicated source function). The RiemannSolverNonLinear change is mostly only for performance.

Note that this change breaks most applications which do have the old source() signature. There was also no much different way of doing this, i.e. adding another named compositeSource(Q,gradQ,S) function would not change too much as it also had to be implemented in the user solver.

This commit also contains some tests as it doesn't compile without.