Cleaned up the main.cpp, introducing the exahype namespace.
By default, nothing changes. However, with a compile flag, user applications can take now control over the main function.

This emphasises the spirit of ExaHyPE as an "game engine", where it is a part of the program but not the whole program.
Of course Peano stil will play the Hollywood game once exahype::main() is called.

Next step: Develop a game using the ExaHyPE game engine. Actually it's already in the repository but you don't yet know it ;-)