Toolkit: Finished experiments with template engines; restructured generic ADERDG glue code generation.
Bottom line is that fining a minimal engine (without megabytes of dependencies) in the modern java world is hard. It's even impossible to find out of the box working JARs, one has to fight for hours with Maven and I now spent more than six hours on this, giving up.

Thus we still are on a level of {{var}} text replacements, at least with a class abstraction in the java code. However, in order to simplify the next steps, I moved all the crazy guard logic back to the glue code (Abstract*Solver).