Godunov 3D: Using template information, at least for the source.
Actually, in the FV solvers we made the much more intelligent decision to call the guards directly instead of heavily using templates to generate all kind of versions of an indvidual kernel function (i.e. template<useFlux,useSource,useNCP>...). In principle, I would favour to change the use*... guards to a constexpr attribute or constexpr method so it should be cheap to evaluate them everywhere instead of relying so much on templates.