Toolkit: Expose variable positions and names to the C++ API.
Thus applications can build on top of this even without using the fancy OOP variable interface.

I need this heavily in all Fortran'ish applications as well as in general prefer to write Q[rho] instead of Q.rho(). Also I strongly need the variable names for instance in the plotters.