Bringing the User API (virtual functions) up to the Solver.h and adding documentation.
Thus, we can guarantee a similar API between FV and ADERDG solvers and by inspecting the header files of the ADERDGSolver.h and FiniteVolumeSolver.h it is immediately visible where the API still has differences (for instance in the lack of a patchwise solution update call in the FV solver).

This change should be fully transparent for a user, except of the renamed user guard "useSource" which now consistently is named "useAlgebraicSource". It was a relic of the last coding week that this renaming step has not been performed. The GRMHD application exemplarily carries out this change: It's just a single rename in the user FV solver header.