Runner: Adjust repository geometry ~ hMax
Adjust the peano repositoriy geometry such that the extents
equal a multiple of the size of the coarsest compute cells
in the grid.

Now adaptive mesh refinement will not introduce
empty cells into the grid.
This happens since finer cells might fit
inside of the geometry while the coarser
ones did not.