FV2VTK plotters: Extending classes for adding a vertex plotter.
That is, before this commit we had FiniteVolumes2VTK a cell-centered plotter, plotting the real DOF of a FiniteVolume Solver.

As an exercise to implement this in the efficient FiniteVolume2CarpetHDF5 plotter which implements the popular, compact and widespread CarpetHDF5 format, I will implement a vertex plotter for the Finite Volume VTK plotter in the next commit.