FiniteVolume2VTK vertex-centered plotting: Finished a first approach.
Unfortunately, for the interpolation I could not take the ghost layers into account as they are not properly available at plotting time.

For instance, in t=0 they are all 0.0 while at later times the corner values are not present (this is due to our FV schemes we actually use).

While theoretically I'm not even sure whether this kind of interpolation makes any sense, at least now I can plot *something* at the vertices which I urgently need in my CarpetHDF5 plotter. The portation of the code to the plotter is something for the next step.

In principle, this interpolation converges to the correct value with decreasing grid spacings.