Plotter: Adding slicing classes.
They abstract the way we define whether a patch shall be plotted and also allow individual finer-grained slicing inside a patch. The CartesianSlicer is a naive attemp to slice planes parallel to the xy, xz or yz plane or the same with axes parallel to x, y or z. This could easily be extended to a way to slice any kind of lower dimensional surface.

Next step is to integrate them into the VTK plotters.