Plotters: Finished the new slicing, implemented for FV2VTK.
So now we really can plot 2D and 1D stripes of patches/cells. That is, on a per-patch basis. Haven't tested too much the RegionSlicing, but the CartesianSlicer is exactly what I want, for 2D, at least for the start.

Based on this I could do really pointwise slicing for ADERDG, at least with the CarpetHDF5 plotter.