LimitingADERDGSolver: Implemented obligatory but useless virtual functions.
This is for the API update when Tobias wanted all virtual functions of our Abstract*Solvers to be virtual right in the Solver. Okay, actually he wanted them to be virtual only in the ADERDG/FV Solver and it was my proposal to have them already in the Solver to reduce duplication. Maybe the ADERDG/FV Solvers should have a common ancestor (VirtualSolver or so) and that's it. Anyway. With our glue code, these functions are never called. I think the current API with the "dead code" virtual implementations in the LimitingADERDGSolver.h is bad.