Misc/ClusterConfig for LOEWE. It's missing GSL. We have to provide it.
That's ugly. I don't want to ship GSL with applications which need GSL. I'll probably write a download-and-install script and put it to ExternalDependencies.

In principle, we have a similar case with TBB which also should be shipped in case of strange clusters or bugs arising due to old versions. There should be version checks in ExaHyPE and the possibility to quickly obtain an own copy by downloading and compiling it from upstream.

Note to myself: Downloading and compiling should be two seperate steps as so many clusters don't have a direct internet connection.