Reintroduce limiter status spreading. Regenerate glue code.
The mapping is important for scenarios
where at the finest level an irregular limiter
domain change occurs. The limiter status
is then restricted to all the parents.
And then spread on each level of the
space tree using the LimiterStatusSpreading
mapping.
This ensures that we do not erase
certain flagged cells during the
consequent mesh refinement iterations.