Plotters::CarpetHDF5: Fully implementing the dimension reduction, producing correct slices.
Now the files can be read by our tools we have in our group and elsewhere. As for instance 2D slices of 3D data are not only the patches around a plane but really the points on the plane, data are again even smaller and easier to handle, for instance for 1D data.