BuildScripts: Adding batch-compile, env-agnostic oot compile kickoff.
This script does not rely on the ENV but instead allows passing the
variables by command line. Thus we are more flexible to actually batch
the compilation of several versions of a code for instance with SLURM.
Otherwise, there's no big difference, except the included handling of
the cluster configs. Which is not a big deal.