GRMHD: Neutral paths. Iboga clusterconfig: neutral exports.
So the first thing about GRMHD is: No dependencies on other applications
in order to avoid out of tree struggles. Instead, we only rely on
ExternalLibraries for the HDF5 include path.

Second, a similiar spirit for the iboga-gcc-tbb: Don't enforce some
specific build configuration but let the user still choose. Don't know
yet if it's a good idea to loosen the connection between build specific
settings or to strengthen it. We'll see. There's nobody using this
cluster config system anyway except me.