Adding new plotter "FlashHDF5".
Code compiles but is completely untested. When you use the new plotter it will crash with absolute certainty.

See FlashHDF5Writer.h for a description and the features it provides. In short words, it uses HDF5 extendible datasets and chunking to create only a few datasets per timestep, saving a lot of overhead especially for an output mode where there is only one physical field per file.