FlashHDF5 output format: Finished and running.
Fixed some bugs. Have not checked the validity of the outputted datas but at a first glance it looks good and there is no more crash at runtime.

This commit also contains some Slicer::toString improvements.

It also contains the GRMHD specfile as the demonstrator with which I tested the FlashHDF5 file format.