{Carpet,Flash}HDF5 plotters: Degrade gracefully if no HDF5 libs built in.
Thus the plotters behave similar to the PeanoHDF5 format which also does not stop the program when HDF5 is missing but just spills a ton of errors whenever it shall plot. Thus the careful user notices that he forgot to compile with HDF5 libraries.

For the careless user I included the possibility to force an abort when he sets the environmental variable EXAHYPE_STRICT. This is helpful when you submit a large job and a proper output is crucial, ie. you don't want to burn CPU hours without any usable output. I think this is a good tactic we can adopt to have ExaHyPE behave nicely (principle of least surprise) in case of incorrect use.