Benchmarks/hamilton/GRMHD: Purifying the application for JM.
The benchmark code for GRMHD contains now only the AlfenWave ID and a single plotter that writes out the error norms. All the overhead code dynamically managing initial data or boundary conditions are removed. Any potential external dependencies are removed. Any fancy plotters are removed. This is now really the minimum.

Note that our GRMHD application, as well as this one here, contains already a Cons2Prim C++-Variant so we avoid the intel compiler bug.

If you cannot run this application, let me know. This is the easiest way we can work with GRMHD.