Toolkit, changing the "Variables" generation.
Namespace members now are constexpr's which is neccessary in order to use the header file in several cpp files without problems at linking.

The class members are currently not constexpr's but "const" instead. This is because the f***ing C++ compiler and me did not get friends. We should also convert them to constexpr's. Which is probably not too hard.