Adding new PDE GRMHD_cpp.
This first commit contains an early version (aka draft) of the PDE system written in a smallish C++ tensor library with focus on ExaHyPE's split of work, i.e. all loops are present in the formulation. The formulation hence is very similiar to the CCZ4 formulation where all tensor contractions are written out as loops.

These C++ files compile as standalone but are untested and contain a number of question marks which need to be resolved.

The integration as ExaHyPE application is straightforward and subject to upcoming commits. Testing with some benchmarks comes afterwards.