GRMHD_cpp: Added an ExaHyPE specfile.
The application compiles, however remains untested. I want to test the routines seperately distinct from ExaHyPE with the old Fortran GRMHD routines. If they provide the same result up to machine accuracy, then we have a new formulation which is C++ only and hopefully saves us all the multithreading headache. It should also be straight forward to start with vectorization form this formulation.