GRMHD_cpp/PDE: Just noticed that the C2P is buggy.
That means the Fortran GRMHD C2P operation which creates errors of O(10%) when running chained: P1->C2P->P2C->P2 and error:=abs(P2-P1) is way too big. Have to dig in deeper tomorrow to understand whether this is a systematic problem in the domain of GRMHD or the code is wrong. This would mean I found an error in the Trento Fortran GRMHD formulation.