GRMHD_cpp: Testing the integration of the PDE into ExaHyPE.
Currently for some reasons a weird error at the ID appears, but this can be resolved hopefully soon. Then we even can see whether the C++ code is faster than the Fortran one ;-)

We should also have some tests with Fortran and we then can couple sophisticated equation of state codes developed in our group into this code. Which is actually something new.