GRMHD: Fixing some minor ID bug. Switching again to C2P in C (Dominic).
Result: As always, the C2P helper routines do not play with Fortran. This was the reason at the first place why I started with a pure C++ GRMHD formulation. It is at least convincing that the Fortran GRMHD application is currently (!) not working. I will switch back to GRMHD_cpp and try to debug it.